refactor(providers): split Anthropic Messages adapter
